+++ evo-conway-life-pixels/src/main.rs
@@ -10,8 +10,8 @@ const MUTATION_ODDS: f64 = 0.0;
 fn main() {
     animate(|window_size| {
         EvoConwayWorld::new(
-            (window_size.width / CELL_PIXEL_WIDTH) as usize,
-            (window_size.height / CELL_PIXEL_WIDTH) as usize,
+            window_size.width / CELL_PIXEL_WIDTH,
+            window_size.height / CELL_PIXEL_WIDTH,
             Random::new(),
         )
     });
@@ -24,13 +24,13 @@ pub struct EvoConwayWorld {
 }
 
 impl EvoConwayWorld {
-    pub fn new(width: usize, height: usize, rand: Random) -> Self {
+    pub fn new(width: u32, height: u32, rand: Random) -> Self {
         let mut result = Self::new_empty(width, height, rand);
         result.add_random_life();
         result
     }
 
-    fn new_empty(width: usize, height: usize, rand: Random) -> Self {
+    fn new_empty(width: u32, height: u32, rand: Random) -> Self {
         assert!(width > 0 && height > 0);
         Self {
             grid: WorldGrid::new(width, height),
@@ -51,11 +51,11 @@ impl EvoConwayWorld {
 }
 
 impl World for EvoConwayWorld {
-    fn width(&self) -> usize {
+    fn width(&self) -> u32 {
         self.grid.width()
     }
 
-    fn height(&self) -> usize {
+    fn height(&self) -> u32 {
         self.grid.height()
     }
